Remove old files.
